# apps/api/openapi/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/users/{id}": {
            "get": {
                "operationId": "UsersController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseUserDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ]
            },
            "patch": {
                "operationId": "UsersController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseUserDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ]
            },
            "delete": {
                "operationId": "UsersController_removeUser",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseUserDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ]
            }
        },
        "/users": {
            "get": {
                "operationId": "UsersController_findAll",
                "parameters": [],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseUserDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ]
            },
            "post": {
                "operationId": "UsersController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDto"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseUserDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ]
            }
        },
...
        "/assistant/user/{id}": {
            "get": {
                "operationId": "AssistantController_findByUserId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseAssistantDto"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "assistant"
                ]
            }
        },
...
        "/doctor/user/{id}": {
            "get": {
                "operationId": "DoctorController_findByUserId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseDoctorDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "doctor"
                ]
            }
        },
...
        "/patient/user/{id}": {
            "get": {
                "operationId": "PatientController_findByUserId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponsePatientDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "patient"
                ]
            }
        },
...
    },
    "info": {
        "title": "MEDIGO API",
        "description": "API Endopoints for MEDIGO",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
            "ResponseUserDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "role": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "firstName",
                    "lastName",
                    "email",
                    "status",
                    "role"
                ]
            },
            "ResponseUserPatientDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "firstName",
                    "lastName",
                    "email"
                ]
            },
            "UpdateUserDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "status"
                ]
            },
            "CreateUserDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                }
            },
            "ResponseSpecialityDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "description"
                ]
            },
            "ResponseDoctorDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "speciality": {
                        "$ref": "#/components/schemas/ResponseSpecialityDto"
                    },
                    "user": {
                        "$ref": "#/components/schemas/ResponseUserPatientDto"
                    }
                },
                "required": [
                    "id",
                    "phone",
                    "speciality",
                    "user"
                ]
            },
            "ResponseAssistantDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/ResponseDoctorDto"
                    },
                    "user": {
                        "$ref": "#/components/schemas/ResponseUserPatientDto"
                    }
                },
                "required": [
                    "id",
                    "doctor",
                    "user"
                ]
            },
...
            "ResponsePatientDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "address": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/ResponseUserPatientDto"
                    }
                },
                "required": [
                    "id",
                    "address",
                    "phone",
                    "user"
                ]
            },
...
        }
    }
}